M3-1434: add tags to Domains list view (#4065)
* M3-1434: add tags to Domains list view

* M3-1434: fix tags linking, combine renderTags function, fix linting

* M3-1434: remove event bubbling from showMore
